# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api covid19 Documentation",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/v1/auth/sign-up": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "Register",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "fname",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lname",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/auth/sign-in": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Login",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/schools": {
            "post": {
                "tags": [
                    "School"
                ],
                "summary": "Creating new School",
                "description": "ECreating new School",
                "operationId": "d3274fb51a26858899259672034054e3",
                "requestBody": {
                    "description": "School Form create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nom.",
                                        "type": "string"
                                    },
                                    "immatriculation": {
                                        "description": "Immatriculation.",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Description",
                                        "type": "string"
                                    },
                                    "devise_fr": {
                                        "description": "devise en Français",
                                        "type": "string"
                                    },
                                    "devise_en": {
                                        "description": "Devise en Anglais.",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "succsess response"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            }
        },
        "/api/v1/school/{id}": {
            "get": {
                "tags": [
                    "School"
                ],
                "summary": "Details school",
                "description": "Show details school",
                "operationId": "5bba81796205d6bf7068367ff99b371f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "School ID identification",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succsess response"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "School"
                ],
                "description": "delete school ",
                "operationId": "6ea6cdd43c96ce1d0de1afc285f6ef09",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "School ID to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "School deleted"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            }
        },
        "/api/v1/schools/{id}": {
            "put": {
                "tags": [
                    "School"
                ],
                "summary": "Updating school information",
                "description": "Upadting school information",
                "operationId": "fe79db5d80863499413c1ab2d8619a38",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID school identification",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Form school updating",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name .",
                                        "type": "string"
                                    },
                                    "immatriculation": {
                                        "description": "authorization id",
                                        "type": "string"
                                    },
                                    "devise_fr": {
                                        "description": "Devise en Français",
                                        "type": "string"
                                    },
                                    "devise_en": {
                                        "description": "Devise en Anglais",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "School description",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "succsess response"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            }
        },
        "/api/v1/schools/users": {
            "get": {
                "tags": [
                    "School"
                ],
                "summary": "return user schools list",
                "operationId": "d0ed6ef49cd26c5eb476e0a1aec65237",
                "responses": {
                    "200": {
                        "description": "return School collection"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}